very wip splitting into files
